pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neural-net"
version = "0.1.0"
description = "Neural Network implementation with MLP, optimizers, and training utilities"
authors = [
    {name = "Santiago Tobio", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "cupy-cuda12x>=12.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "ruff>=0.1.0",
    "mypy>=1.4.0",
    "ipython>=8.12.0",
    "jupyter>=1.0.0",
]

[project.scripts]
neural-net-train = "scripts.train:main"
neural-net-eval = "scripts.evaluate:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["neural_net*", "experiments*", "scripts*"]
exclude = ["tests*", "archive*", "legacy*"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
# Enable pycodestyle (`E`), Pyflakes (`F`), isort (`I`), and more
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (Black handles this)
    "B008",  # do not perform function call in argument defaults
]
line-length = 88
target-version = "py310"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start lenient, tighten later
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

# Ignore missing imports for external libraries without stubs
[[tool.mypy.overrides]]
module = [
    "cupy.*",
    "matplotlib.*",
    "seaborn.*",
    "tqdm.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=neural_net --cov-report=term-missing --cov-report=html"
